pciepsu: Fix MISRA 12.2 violations
MISRA Rule 12.2 - The right hand operand of a shift operator shall lie
in the range zero to one less than the width in bits

Fixing MISRA violation 12.2 by casting to proper width when doing shift ops.


Acked-for-series: Bharat Kumar Gogada <bharat.kumar.gogada@xilinx.com>

+++ XilinxProcessorIPLib/drivers/pciepsu/src/xpciepsu_ep.c
@@ -80,7 +80,7 @@ void XPciePsu_EP_BridgeInitialize(XPciePsu *PciePsuPtr)
 	Val = XPciePsu_ReadReg(PciePsuPtr->Config.BrigReg,
 			XPCIEPSU_E_BREG_CONTROL);
 	Val &= ~(BREG_SIZE_MASK | BREG_ENABLE_FORCE);
-	Val |= (BREG_SIZE << BREG_SIZE_SHIFT);
+	Val |= ((u32)BREG_SIZE) << BREG_SIZE_SHIFT;
 	Val |= BREG_ENABLE;
 	XPciePsu_WriteReg(PciePsuPtr->Config.BrigReg,
 			XPCIEPSU_E_BREG_CONTROL, Val);
